rename get.ts to node.ts
