+++ src/api/tree/query/node.ts
@@ -3,7 +3,7 @@ import { forceLeadingSlash, sortFoldersFirst } from "../utils";
 import type { QueryFunctionContext } from "@tanstack/react-query";
 import { TreeListSchema } from "../schema";
 import { apiFactory } from "../../utils";
-import { treeKeys } from "../";
+import { treeKeys } from "..";
 import { useApiKey } from "../../../util/store/apiKey";
 import { useNamespace } from "../../../util/store/namespace";
 import { useQuery } from "@tanstack/react-query";
